CommonsLinkChecker uses commons db for files now
Formerly checked with HTTP request (not secure).
Other commons namespaces than File are not covered anymore (but file is
the common case so it maybe is not too bad until we find out how to check
the others).
Also not too nice: MW_PHPUNIT_TEST query for name of database ...

Bug: T101468
Bug: T101469
Change-Id: I54e8850cde12da86a6c474d6e94f04bdaf0c9e4d