Store entity ID in check result
The API will support checking for all violations of a given claim (by
claim ID), and it will return check results grouped by entity ID, so it
needs to know the entity ID of each check result. To avoid having to
parse the entity ID out of the claim ID twice, we store it in the check
result. This means that checkConstraint()’s third parameter, $entity, is
no longer optional, which in turn requires many small changes to the
tests, even though most of the tests don’t actually use the entity.

Change-Id: I06d1256dc9c09e390aaf1367d40b5b1890c2142f