Import constraints from statements on properties
Constraints are modeled as statements on a property, using the property
“property constraint” (P2302 on Wikidata) and different items for the
different kinds of constraints. (Constraint-specific properties like
“format as a regular expression” (P1793 on Wikidata) are not used in
top-level statements, and should instead be used as qualifier in a
“property constraint: format constraint” statement.)

Constraints from statements are still entered into the database just
like imported constraints; to do this, we register a hook that runs each
time a property changes. To avoid blocking these changes, the actual
work of extracting the constraints is moved to a job, and the hook only
checks if the job needs to be run and then enqueues the job.

In this first version, only constraints without parameters are
supported. Later commits will add support for reading parameters from
qualifiers on the constraint statements.

All IDs used in constraints are configurable, and this commit includes a
command to import them from Wikidata, and a script to export their IDs
to the configuration file; see the “Data import” section of the
README.md file for details.

The import job is hidden behind a feature flag which defaults to false;
if the WBQualityConstraintsImportConstraintStatements configuration
option is not set to true in LocalSettings.php, no constraints from
statements are added into the database, and the rest of this change
remains without any effect.

Bug: T102760
Bug: T160505
Depends-on: I0461e59da2a8fa6681e3b1fcdfc38bfed7f3ac32
Change-Id: I03cd98769231895c0aa30f4b1fc03f6aa3ad1c73