Add two examples for the API
One example for the id parameter, and one for the claimid parameter. The
item IDs in the id example should hopefully exist on most Wikibase
installations; the statement ID in the claimid example is unfortunately
Wikidata-specific (it’s the ID of Douglas Adams’ “religion: atheism”
statement).

An example for the constraintid parameter can be added once the
migration to reading constraints from statements is done (which will
make the constraintid a statement ID instead of a UUID).

Change-Id: Ia8c9c1ecf02732d2d56c2d96f87967c15bb3d62c