Encode empty constraint params as {} instead of []
This doesn’t make any difference when loading the constraints, but it
makes tests that directly look at the database and compare inserted
values a bit more consistent (otherwise they need to expect '[]' for
empty parameters and '{...}' for nonempty ones).

Also expands an existing test to test for this change by making it
insert a constraint with no parameters.

Change-Id: I1a00edf77173ee7cbcecf7054313fb389366c291