Add margin-inline-start to complement margin-left
margin-left is incorrect in right-to-left languages, where we need the
margin on the right side. Ideally, we’d replace it with
margin-inline-start, which does what we need in all directions, but
since that property is only supported in Firefox according to MDN [1],
we keep margin-left as well. (In Firefox, having both margin-left and
margin-inline-start doesn’t hurt – they’re not additive if they both
refer to the same margin.)

[1]: https://developer.mozilla.org/en-US/docs/Web/CSS/margin-inline-start#Browser_compatibility

Change-Id: Ia1700a31030001152f5672f34a0cc5ed6a7ea957