Improve violation messages for Type and Value type
All the important information for the user is now encoded into the
message itself, instead of just referring to the parameters: the
property, the entity, and the classes. To make this possible, the
TypeCheckerHelper and the ConstraintReportFactory now get an
EntityIdFormatter injected (usually an EntityIdHtmlLinkFormatter, but
can also be a PlainEntityIdFormatter for tests).

Languages can choose to inline any number of classes into the message
text, or fall back to an HTML list if there are too many classes. (The
English message has special cases for one or two classes, and uses the
HTML list if there are three or more classes.)

Bug: T164354
Change-Id: I1aa83f5e03972afbe36b46e41cb081e2a9258202