Let RangeChecker fall back to {minimum,maximum}_quantity
RangeChecker assumes that if the data type is "time", the range is
defined by the parameters minimum_time and maximum_time. However, the
ConstraintsFromTemplates import script always produces ranges with
parameters minimum_quantity and maximum_quantity, regardless of the type
of the property. To fix the resulting “incorrect parameters” violation,
RangeChecker now falls back to checking for _quantity parameters if the
_time parameters are not found.

The tests do not need to be adjusted because the tests that check for
incorrect parameters supply one _quantity and one _time parameter, which
is still rejected – the RangeChecker will only accept both _quantity or
both _time parameters, and only if no parameters of the other kind are
present.

Bug: T164087
Change-Id: I06c5680cbe8b9040e62f04b0405d552637b1d218