Replace "now" with current time before comparing
The special string "now" corresponds to the current time, so we turn it
into the current time formatted like a TimeValue.

(Note that due to T164279, the string will not be compared correctly –
this will be fixed in a separate change.)

Also adds a test, but due to the bug mentioned above, the min date for
the test has to be formatted differently so that the string comparison
works – a later change should clean that up.

Bug: T90107
Change-Id: I3fb31596c9c7a7447caa6bbfdb911b553da35daa