Remove DataValue objects from test parameters
Outside of the tests, the parameters are parsed JSON objects from the
database – it doesn’t make sense to put objects in them: if the checkers
need objects instead of strings, they have to parse the strings.

This causes one test to fail (the checker doesn’t actually parse yet),
so we skip the test for now.

Bug: T164279
Change-Id: Ia9c67d90fead25d87ab108b781d49fb44d4683bc