Fix mobile frontend detection
Thiemo pointed out that the previous version was flawed, since the
condition would succeed if anything caused any mobile frontend code to
be loaded, regardless of whether it was active or not. Instead, we can
check for the configuration variable wbMFMode, which is null on desktop
mode and non-null on mobile (I got 'stable' on Wikidata, presumably
there are other modes as well).

Change-Id: I7a64dfa0dfc5cf5fe51e3bd0ac38b0cff54c4a83