Streamline HTML generation code in SpecialConstraintReport
"Html::openElement + raw HTML + Html::closeElement" really is the same as
Html::rawElement.

Warning: I'm not sure if it's correct to use Html::rawElement in all these
cases! It might be that some need HTML escaping and must use Html::element
because of this. But if such cases exist they are not introduced by this bug,
and should be fixed in a later one.

Change-Id: I53dbef8ae2e8aa1cdb0d90650499d3cf15b3f536