Map item and property IDs from parameters to uppercase
In the constraint templates on property talk pages, items and properties
are linked using the Q and P templates of Wikidata. Since the first
letter of a template name is always mapped to uppercase, it is also
possible to specify the templates as e. g. {{p|2676}} instead of
{{P|2676}}, and some constraints on Wikidata, e. g. the Qualifiers
constraint on P2363, do this. In this case, the ConstraintsFromTemplates
script preserves the lowercase p, and the parameters contain a parameter
p2676 instead of P2676. When a checker, such as the
MandatoryQualifiersChecker, then compares the IDs from the parameters
with the proper serializations of actual entity IDs, it erroneously
reports a mismatch. (The particular case described as example here is
the cause of bug T164731, which this commit fixes.)

In effect, we currently have to assume that item and property IDs in the
constraints table may use any capitalization, whereas the serializations
returned by the ItemId and PropertyId classes are always uppercase. To
compare them correctly, we must therefore first map the parameter IDs to
uppercase.

Many checkers that operate on entity IDs use a ConnectionCheckerHelper,
so we can move the mapping for those checkers into the helper. The other
checkers have to be updated to map to uppercase themselves.

This commit can hopefully be reverted once constraints are imported
exclusively from statements on properties.

Bug: T164731
Change-Id: I76d961030d39f452fc06c08e58649952192424aa