Import constraint parameters from qualifiers
The updater job is kept very simple: it just serializes all qualifiers
and adds them all to the parameters. Interpreting those qualifiers
(starting with mapping their property IDs to parameter meanings) is then
left to the individual checkers, and done in a separate commit.

A test is added to test the parameter extraction method; for a full test
of importing a complete constraint with parameters, we have to wait
until we have config variables for some constraint types that take
parameters, as well as for those parameters. This will happen in a later
commit.

Change-Id: I6e5428187fc9d0c3137ca1fb0b20c18567c4698c