Fix MAX_ENTITIES check
There was an off-by-one error – we need to increment the counter before
comparing it. Now, if the limit is 1, on the first call the counter will
be incremented from 0 to 1, and the comparison 1 > 1 will fail, and then
on the second call the counter will be incremented to 2 and the
comparison will succeed, so the number of full method executions will
have been 1, matching the limit.

Change-Id: Iffe364a9f066b118414c3b25986212191914aa25