Add expectation for number of calls
If the cycle detection in TypeCheckerHelper doesn’t work, it will now
cause a test failure instead of a test that runs forever.

Includes a refactoring to turn a member variable initialized in setUp()
into a getter function – expectations on mocks created in setUp() have
no effect, so it’s better to create the mocks in the context of the test
method.

Change-Id: I231421a52bff138e17d60696b66f090cd8d0ea22