Return found entity in ConnectionCheckerHelper::hasClaim
hasClaim() now returns the value of the claim that was found, or null if
nothing was found, instead of a boolean. Using hasClaim() in an if
clause continues to be valid, since null converts to false and any
non-null value converts to true, so checkers that use the method do not
need to be updated. However, the test assumptions need updates, which is
also a good opportunity to replace assertEquals( true, … ) with
assertTrue( … ) and similar changes.

Also includes various other improvements to the hasClaim() code, since I
was already looking over it.

This will allow the “Conflicts with” checker to include the particular
conflicting statement in the violation message.

Change-Id: Ib64ecc51d8618b0c69bff88dc9bd726db3072e81