Improve violation messages for Conflicts with
The messages now get two to three parameters, with the current property,
the conflicting property, and (if applicable) the conflicting item.

The messages previously already got a single parameter, containing the
name of the checker, but the parameter isn’t used in any language, so
changing does not break any messages.

Bug: T164354
Change-Id: I0ef5589048700779205fdee539c54d5585a71992