Improve violation messages for Qualifiers and Mandatory qualifiers
The message now includes all the information necessary to understand the
violation, including a bit that was not previously available in the
parameters: the property that causes the violation because the qualifier
is not permitted (Qualifiers) or missing (Mandatory qualifiers).

The special case of no permitted qualifiers at all is turned into a
separate message, since it differs enough from the message for one or
more qualifiers that accommodating them both in a single message would
require too much PLURAL magic.

Bug: T164354
Change-Id: Ic876f160b8e04912f3597a64282179686c78fb32