Improve violation messages for Range and Diff within range
The messages now get four / six parameters: the property/properties,
value(s), and range.

In DiffWithinRangeChecker, the variables relating to the linked property
($statement, $mainSnak) have to be renamed ($other…) so that the
original variables are still available when assembling the message. This
also fixes an unrelated bug: in the CheckResult instantiation,
$statement was always meant to refer to the checkConstraint() parameter,
not to the iteration variable (now $otherStatement). The result of this
bug was that the API returned the constraint under the wrong statement
ID and the user script consequently displayed the violation under the
linked property instead of the one that had the constraint.

Bug: T164354
Change-Id: I0d36bc219c18418f5ab894297961b0f9ee1a4e65