Make MAX_ENTITIES configurable and raise default
The limit should be configurable, since different projects with
different class hierarchy conventions will need different values. For
Wikidata, the old limit of 50 was rather low (it is common to have
several, partially redundant superclass: city, big city, metropolis, …),
so the default is raised as well.

Hopefully, we will soon be able to use a SPARQL-based type checker
instead; in that case, this limit will probably become the point at
which the checker switches from direct supertype traversal to SPARQL
checking, and a lower value will be more appropriate.

The DefaultConfig trait for tests reduces the limit to make the tests
run a bit faster.

Bug: T164948
Change-Id: Ic5bf0f5453121f8b77e8616ba717b22fea8192e9