EntityIdFormatter → ConstraintParameterRenderer
All services that need to format entity IDs for their parameters now get
a ConstraintParameterRenderer instead of an EntityIdFormatter. This
removes some code duplication, because ConstraintParameterRenderer
already knows how to format an HTML list, or how to do the
try-parse-catch-htmlentities dance for an entity ID string, so the
corresponding copied code in the individual checkers / helpers can be
removed.

ConstraintReportFactory’s EntityIdFormatter is no longer necessary and
therefore removed.

All test cases are also updated to inject a ConstraintParameterRenderer
into the classes they test instead of just an EntityIdFormatter.

Change-Id: I86adfb5afeef0836fc7b489fc1b5eaa9733acdff