Don’t include time of day when interpreting 'now'
The *rationale* for this change is that Wikibase doesn’t support sub-day
precision yet, so there’s no need to compare the time of day when
checking the range. If 'now' is the maximum, this changes nothing; if
'now' is the minimum, the current day is now accepted whereas it was
previously rejected, but there are no constraints on Wikidata where
'now' is the minimum.

The *motivation* for this change is I0d36bc219c: Wikibase’s default
value formatters format date strings with non-zero time of day as plain
ISO-8601 string instead of as localized text like “May 16, 2017”, so
changing this precision makes the value look nicer in the message.

Change-Id: I950746de18b99ca3a3cd8f36f495883ce642fc93