Fix README.md links
The links were broken on GitHub because according to CommonMark [1],
which GitHub Flavored Markdown [2] is based on,

> [Whitespace] is not allowed between the link text and the following
> parenthesis[.]

The specification demonstrates this in example 480 [3].

The minimal change to fix the links would be to just remove the line
breaks between ) and [, but using reference links [4] makes the Markdown
much more readable.

An archive.is snapshot from 2015 [5] shows the links being rendered
correctly, even though according to Git history the problematic line
breaks were present ever since the links were first added in dfd93b0e,
so I suspect that the links were broken by GitHub’s move to base GitHub
Flavored Markdown on CommonMark [6], and that this README.md was one of
the few affected documents that could not be migrated (because, unlike
issues and comments, they were part of Git history).

[1]: http://spec.commonmark.org/
[2]: https://github.github.com/gfm/
[3]: http://spec.commonmark.org/0.27/#example-480
[4]: http://spec.commonmark.org/0.27/#reference-link
[5]: https://archive.is/Sp3lb
[6]: https://githubengineering.com/a-formal-spec-for-github-markdown/

Change-Id: Ic5956833d4f555ac528b32eb293b0bf7e3d4acd4