Sort constraint results by constraint type
If two constraint results do not differ in anything else, sort them by
the constraint type, to ensure that constraint results for the same
constraint type are grouped together. This especially matters for the
output of the constraint-displaying user script, since the constraints
shown in a single report pop-up already share the other aspects that the
sorting takes into account (status, property, statement).

Properties with multiple constraints of the same type are not common,
but not unheard of either; constraints like “Item” (item requires claim)
or “Conflicts with” can be specified with completely different
parameters. The following query finds such properties:

    SELECT DISTINCT ?property ?propertyLabel ?constraintLabel WHERE {
      ?property a wikibase:Property;
                p:P2302 ?statement1, ?statement2.
      ?constraint ^ps:P2302 ?statement1, ?statement2.
      FILTER(?statement1 != ?statement2).
      SERVICE wikibase:label { bd:serviceParam wikibase:language "en". }
    }

Change-Id: I2dfabdd725ad290a9b04390f94d0bb9d0b932110