Add test for extracting constraint parameters
This is a test for functionality introduced in commit 0c03ee2a24 (change
I6e5428187f), but it uses configuration variables added only in commit
8e9438ea7c (change Id47731c99b), so it is added in a separate commit.

testImportConstraintsForProperty and testRun / P2.json are not updated
because the added benefit seems small compared to the large amount of
extra code that would be needed.

Also fixes some confusing comments in other tests.

Change-Id: I88f03425a3bb9d6dd1900b73579faf36b94fdfeb