Extract ResultAssertions trait in tests
The main methods this trait offers are assertCompliance(),
assertViolation() and assertTodo(). Their main task is to compare the
check result status to the appropriate constant, which reduces copied
code in all the checker tests.

Since slightly reduced code repetition is not very exciting, some new
features are also introduced:

- assertCompliance() includes the result message in the assertion
  failure message, which should make it easier to understand test
  failures.
- assertViolation() optionally also takes an expected message and
  asserts that the actual result message matches the expected one. This
  uncovered the message bugs fixed in change I6a5bb8d3a3.

Additionally, since the FormatChecker tests already included the
information whether a check is expected to succeed or not even though
the check is not yet implemented, two aliases for assertTodo() are also
available, assertTodoCompliance() and assertTodoViolation(). This should
make it very easy to adapt the tests once the constraint is implemented.

assertViolation(), when comparing the result message, expects each
parameter in the raw message to be replaced with a nonempty string, so
some mocks need to be updated to provide getConstraintTypeName()
(included in some messages), and the ValueFormatter mock needs to be
updated to return a nonempty string (included in the range messages).

Change-Id: Ia1b1e1515440759d2df6e2bef04b072395afe7ef