Add some support for somevalue and novalue
A new class is added, ItemIdSnakValue, which wraps an ItemId or the
special values 'unknown value' or 'no value', which for some constraints
need to be treated like ordinary values.
ConnectionCheckerHelper::findStatement is updated to accept a list of
such values (support for an array of strings, or a single string, will
probably be removed eventually), and returns such a value to indicate
the found statement. ConstraintParameterRenderer is taught to render the
new kind of value.

Most call sites of ConnectionCheckerHelper::findStatement only check if
the returned value is null or not, and do not need to be updated;
ConflictsWithChecker is updated to pass the returned value into the
correct ConstraintParameterRenderer method.

This is mostly preparation for importing certain constraints (Item,
Target required claim, One of, …) from statements, but specifying the
strings 'somevalue' or 'novalue' in a constraint template parameter is
now also supported.

Two tests for ConnectionCheckerHelper are added, to test that
'somevalue' and 'novalue' work as expected. They do not yet use
ItemIdSnakValue; once ConnectionCheckerHelper is updated to only use
ItemIdSnakValue, all tests for it will have to be updated anyways.

Bug: T102759
Change-Id: I40cc1aec0c887f6111fe169929f5bf6e1a907bc3