Import Target required claim constraints from statements
TargetRequiredClaimChecker gets a ConstraintStatementParameterParser
injected, which it uses to parse the property and items parameters. A
config variable for the constraint type item is added, and
ConstraintReportFactory aliases it to the “Target required claim”
constraint type.

The test for a missing property parameter is removed, since it now
throws an exception and is covered by the dedicated
ConstraintStatementParameterParser tests. A new test is added which
checks that a constraint from a statement also works.

Bug: T167109
Change-Id: I047e27edfba90a3b68d05bc5020b9ecc2c4a91d1