Add ConstraintStatementParameterParser helper class
This class will contain all logic to parse constraint parameters, both
from statements and (until we remove support for that) from templates.
We will gradually migrate all checkers to use
ConstraintStatementParameterParser, and then remove the current
ConstraintParameterParser and rename ConstraintStatementParameterParser
to it.

We will start out with the Type and Value type checkers, so the helper
can currently parse “class” and “relation” parameters, and configuration
variables for the required entity IDs have been added to extension.json.

If the parameters are invalid, the helper throws an exception of the new
type ConstraintParameterException. The exception is caught in the
DelegatingConstraintChecker, so that the individual checkers don’t have
to handle this case – after all, the only CheckResult parameter that
DelegatingConstraintChecker doesn’t know is the constraint parameters,
which in this case can be set to [] since the parameters are invalid.

The ConstraintParameters test trait gains a method to instantiate the
new helper, and a test class for the new helper is also added.

Change-Id: Iefc5f6d9e21a1dc996f51bb84853e40e4fae545e