Import Type and Value type constraints from statements
ConstraintReportFactory, TypeChecker and ValueTypeChecker all get a
ConstraintStatementParameterParser injected. ConstraintReportFactory
passes the helper to the checkers, which delegate all parameter parsing
to the helper.

Configuration variables for the two constraint type item IDs are added,
adn ConstraintReportFactory aliases them to the “Type” and “Value type”
constraint types.

Tests for the checkers’ behavior with missing parameters are removed,
since those tests now throw an exception (which is only turned into a
CheckResult in DelegatingConstraintChecker, not in the individual
checkers) and are covered by the dedicated tests for
ConstraintStatementParameterParser.

A few new tests for the checkers are added, which use statement-type
parameters instead of template-style ones.

Bug: T167106
Change-Id: Id47731c99b6e8ae10d540315fc68994cf2f0f24e