Import Inverse constraints from statements
InverseChecker gets a ConstraintStatementParameterParser injected, which
it uses to parse the property parameter. A config variable for the
constraint type item is added, and ConstraintReportFactory aliases it to
the “Inverse” constraint type.

The test for a missing property parameter is removed, since it now
throws an exception and is covered by the dedicated
ConstraintStatementParameterParser tests. A new test is added which
checks that a constraint from a statement also works.

Bug: T167107
Change-Id: Idf365b7445af7918208cc56342b8646fb6a13e97