Import Item constraints from statements
ItemChecker gets a ConstraintStatementParameterParser injected, which it
uses to parse the property and items parameters. A config variable for
the constraint type item is added, and ConstraintReportFactory aliases
it to the “Item” constraint type.

The test for a missing property parameter is removed, since it now
throws an exception and is covered by the dedicated
ConstraintStatementParameterParser tests. A new test is added which
checks that a constraint from a statement also works.

Bug: T167108
Change-Id: I2571db085f883bd17b650208c507aeb86b728ba1