Import Mandatory qualifier constraints from statements
In constraint statements, the “mandatory qualifier” constraint only
supports a single property parameter – to add multiple required
qualifiers, multiple constraints should be added [1]. This is a change
from the constraint template, which permitted multiple qualifiers; since
we don’t plan to support both styles of constraints for very long, and
only 16 “mandatory qualifier” constraints on Wikidata specify multiple
properties [2], we remove support for this now, and silently drop the
extra properties in parsePropertyParameterFromTemplate.

The violation message is updated accordingly: the list of missing
qualifiers is replaced with a single parameter for the one missing
qualifier. Since this is a breaking change to the message, the message
key is changed from …-qualifiers to …-qualifier, which also fits the
new, singular interpretation better.

Apart from that issue, this change is analogous to Item (I2571db085f),
Target required claim (I047e27edfb), Conflicts with (I13d95bf523) and
One of (If8792804b3).

[1]: https://www.wikidata.org/w/index.php?title=Help_talk:Property_constraints_portal/Mandatory_qualifiers&oldid=487518545#Modeling_of_multiple_qualifiers_in_constraint_statements
[2]:
    csvawk '$3=="Mandatory qualifiers" { print $4 }' wikidataconstraints.csv |
        sed 's/^"//; s/"$//; s/""/"/g;' |
        jq .property |
        grep -cF ,

Bug: T167416
Change-Id: Ieef2892a3a8eb4ac32d21b3c5cae1001a71c165a