Document @throws and simplify UniqueValueChecker
I forgot to update UniqueValueChecker in I4b90142a8d – it also no longer
needs to catch the SparqlHelperException (DelegatingConstraintChecker
does that now).

Also add @throws documentation tags to the five checkers that can throw
a SparqlHelperException – the exception is already documented in the
general ConstraintChecker interface, but it’s a good idea to repeat it.

Change-Id: I8a481da2b37aa0f928e9ea654e089ba3231925c1