Migrate user script to ResourceLoader gadget
A new ResourceLoader module, wikibase.quality.constraints.gadget, is
added, which includes the old user script code (now gadget.js and in
modules/). Simply loading the module will load the script and trigger
the constraint check.

Since it is trivial to include a CSS file in a ResourceLoader module,
several styles are moved from the script to a new CSS file. This also
lets us drop the margin-inline-start: CSSJanus will turn the margin-left
into margin-right if necessary. (The z-index cannot be moved because it
needs to be directly on the popup element, not on the content element
inside it.)

ResourceLoader modules can define a skip function, which causes the
module to be skipped if the function returns true. We use this to
disable the module on non-entity pages or on the mobile frontend. (Just
to be sure, the script itself also checks these conditions, just as it
still declares its dependencies even though they should already be
loaded.)

A new section in README.md describes how to define the gadget.

The existing wikibase.quality.constraints.ui module remains unchanged,
so the old user script continues to work. However, if both the user
script and the gadget are used, two constraint reports will be shown.

Bug: T168412
Change-Id: I7bef8122e8252c5dd9dc4315c897c589d89895a2