Check constraints on saved statements
When a statement is saved, we check it again, using the
wikibase.statement.saved hook added in I5efdb67c57 and the claimid API
parameter.

A small change to addReportsToStatement is necessary, since in edited
statements, the value is wrapped in a .valueview-instaticmode div; if we
append the button to the surrounding .wikibase-snakview-value instead of
the .valueview-instaticmode div, they appear on separate lines.

Bug: T168627
Change-Id: I2bab15b792fbc1201cbea56b0250b7fad906968f
Depends-On: I5efdb67c57106c64dd2bee0e888ba1bb230886ea