Update QualifiersCheckerTest to use statement parameters
The ConstraintParameters trait gains a method to construct a
'properties' parameter from a list of property ID serializations. The
three “old” tests for QualifiersChecker are updated to use that method
for their parameter, and also refactored to follow the
Arrange-Act-Assert pattern. The separate test method
testQualifiersConstraintWithStatement is removed, since it is now
identical to testQualifiersConstraint.

Change-Id: I3f7cc80153e43bfd4635203cf56b11c7c0b8959e