Add support for importing range parameters
ConstraintStatementParameterParser gains a new method to parse
parameters for a quantity or time range. This is used by the Range and
Diff within range constraint types.

A message that a parameter must be a value or no value, but not unknown
value, is added, and the wbqc-violation-message-range-parameters-needed
message is updated to no longer hard-code the constraint type name,
since it will be used for both Range and Diff within range.

Note that the statement version of the time range parameters does not
yet support 'now', since it is not yet clear how that should be modelled
(see discussion at [1]).

[1]: https://www.wikidata.org/wiki/Help_talk:Property_constraints_portal/Range#special_parameter_.E2.80.9Cnow.E2.80.9D_in_constraint_statements

Change-Id: I14bb590d3eb4ea421dd9d169cc737b895e857c65