Migrate special page hover functionality to CSS
The tooltip element is moved inside the indicator element, so that we
can select for the tooltip inside a hovered indicator. A side effect of
this is that hovering over the tooltip counts as hovering over the
indicator, which means that the tooltip doesn’t vanish when you move the
mouse from the indicator to the tooltip and therefore makes the links
inside the tooltip clickable.

(The migration to CSS can also be achieved without any change to the
DOM, by using the CSS selector `.wbqc-indicator:hover + .wbqc-tooltip`
instead of `.wbqc-indicator:hover .wbqc-tooltip`, but in that case the
side effect is lost and the tooltip is harder to use, since it vanishes
the moment you hover it.)

Change-Id: I91f4de9338e9b4ff6001cc86f164acf15cf28034