Import Diff within range constraints from statements
DiffWithinRangeChecker gets a ConstraintStatementParameterParser
injected, which it uses to parse the range and property parameters. A
config variable for the constraint type item is added, and
ConstraintReportFactory aliases it to the “Diff within range” constraint
type.

RangeCheckerHelper::getDifference() returns a quantity instead of a raw
number, and DiffWithinRangeChecker uses
RangeCheckerHelper::getComparison() to compare that quantity with the
quantities from the range. Units are still ignored, but once T164371 is
fixed, getDifference() can be changed to return the difference in days,
which will allow more fine-grained constraints.

As in Ide78cc0e79, the violation message gets −∞ and ∞ as arguments for
half-open ranges until T168268 adds dedicated messages for those cases.

Some tests for RangeCheckerHelper are removed, since tests for
constraint parameters are now handled by the
ConstraintStatementParameterParser tests. Two new tests for a half-open
difference range (one compliant, one in violation) are added.

Bug: T167414
Change-Id: I36772732a5239cb26dc92fdde479dd2ff73f3331