Support other namespaces in CommonsLinkChecker
Search for the title in the page table instead of the image table, which
contains pages of all namespaces, not just the File: namespace. Since
the page title DB key is not the same as the page title in the
statement, a TitleParser is injected into CommonsLinkChecker (and
ConstraintReportFactory) and used to parse the title, which also gives
us the namespace number we also need for the database query.

commonsLinkIsWellFormed() is slightly relaxed, since titles can contain
colons (and there are images on commons with colons in the file name);
the error case of including the namespace in the value is detected after
the file has been determined to not exist.

Tests for the category and gallery namespaces are added, since those are
supported now. The other “todo” tests are removed. A TitleParser mock is
added (as a trait, so that it can be shared between tests), so that the
tests don’t bear the performance cost of full-featured title parsing.

Bug: T102753
Change-Id: I4a42adf835b0cd139de1feb383fe6703e4f12c54