Add support for importing namespace parameters
ConstraintStatementParameterParser gains a new method to parse a
namespace parameter. This is used by the Commons link constraint type.
(The parseStringParameter() helper method will also be used for parsing
a format parameter for the Format constraint type.)

A message that a parameter must be a string, not another kind of data
value, is added.

Bug: T167417
Change-Id: I2c7a331e4cb7dfc5b1455120000894976a97257b