Import Commons link constraints from statements
CommonsLinkChecker gets a ConstraintStatementParameterParser injected,
which it uses to parse the namespace parameter. A config variable for
the constraint type item is added, and ConstraintReportFactory aliases
it to the “Commons link” constraint type.

The ConstraintParameters test trait gains a method to create a namespace
parameter, and all tests for CommonsLinkChecker are updated to use that
method. This way the tests will continue to work as we remove support
for template parameters.

Bug: T167417
Change-Id: Ibe7a011f5be11fc764d7c24088e76ffeb51caae3