Check constraints on somevalue/novalue statements
All the checkers that actually use the main snak (many, e. g. Item or
Conflicts with, don’t) already check the snak type, so all that needs to
be done is to remove the check in DelegatingConstraintChecker.

(InverseChecker was missing the parameter to the message that is
displayed in this case, so the parameter is added.)

Note that without I8f2783d4bf, the constraint report gadget on these
statements is styled like the “no/unknown value” text (gray and italic),
which is undesirable.

Bug: T103113
Change-Id: I76b02b264d24771b7712fc0c907a1d7c7504ea44
Depends-On: I8f2783d4bf01d0fb083f79d3e29e0e5ed86d0164