Remove incorrect message argument
In I14bb590d3e, I accidentally added the new constraint type name
argument for the wbqc-violation-message-range-parameters-needed message
twice: once at the end, where it belongs (for backwards compatibility),
and once at the beginning, so that it shifted all other arguments (which
is exactly what I wanted to avoid by adding it at the end). Remove the
argument at the beginning to fix the message.

Change-Id: I8e195e87f5ad3e0395916b56eb32f0daa1370db0