Remove constraint_status parsing from checkers
In most cases, this code no longer works after the change from
ConstraintParameterParser to ConstraintStatementParameterParser (the
latter class has no parseSingleParameter() method, and in some cases the
member holding it was renamed from helper to constraintParameterParser).
Instead of fixing it, which would be fairly pointless (the constraint
status is currently ignored, and in constraint statements it is modeled
differently anyways), just remove it. T164254 will add proper constraint
status handling later.

(The broken code is only triggered by constraints which have a
constraint_status parameter, which never happens in the tests.)

Bug: T168629
Change-Id: I38d39a8c342bb76d8ee35109b1680c8225fecb11