Add support for importing format parameters
ConstraintStatementParameterParser gains a new method to parse a format
parameter. This is used by the Format constraint type.

parseFormatParameterFromTemplate also HTML-unescapes the format, which
fixes T164065. It also strips off <code> tags that can be found in some
templates on Wikidata (e. g. on P1553).

(The configuration setting name is deliberately bulky, instead of just
WBQualityConstraintsFormatId – it’s possible that we’ll need a different
kind of format (not regular expression) as part of T102752, and in that
case we’ll want to use a different configuration setting.)

Bug: T164065
Bug: T167418
Change-Id: Iea4cbd6d31fc60512c48312484ea7ceb7dfca944