Import Format constraints from statements
FormatChecker gets a ConstraintStatementParameterParser injected, which
it uses to parse the format/pattern parameter. A config variable for the
constraint type item is added, and ConstraintReportFactory aliases it to
the “Format” constraint type.

The ConstraintParameters test trait gains a method to create a format
parameter, and all tests for FormatChecker are updated to use that
method. This way the tests will continue to work as we remove support
for template parameters.

A test for a missing format parameter is removed, since that is now
covered by the ConstraintStatementParameterParser tests and throws an
exception instead of returning a check result.

Bug: T167418
Change-Id: I1a69a7485cbb8878c633515f693a7248ba9603e0